Moved bunch of logic from scripts.dataset into datacube.index.hl
Logic that deals with conversion of plain metadata docs into `Dataset` objects.
